requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
black==24.8.0
blinker==1.8.2
cachetools==5.5.0
ccimport==0.4.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configargparse==1.7
crcmod==1.7
cryptography==43.0.1
cumm-cu118==0.4.11
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
descartes==1.1.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.14.0
fire==0.6.0
flake8==7.1.1
flask==3.0.3
fonttools==4.53.1
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
idna==3.8
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
lazy-loader==0.4
llvmlite==0.41.1
lyft-dataset-sdk==0.0.8
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.2.0
mmdet3d==1.4.0
mmengine==0.10.4
model-index==0.1.11
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.22.0
nuscenes-devkit==1.1.11
oauthlib==3.2.2
open3d==0.18.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.2
plotly==5.24.1
pluggy==1.5.0
plyfile==1.0.3
portalocker==2.10.1
prompt-toolkit==3.0.47
protobuf==5.28.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pyflakes==3.2.0
pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2023.4
pywavelets==1.4.1
pyyaml==6.0.2
referencing==0.35.1
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.20.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==60.2.0
shapely==1.8.5.post1
six==1.16.0
spconv-cu118==2.3.6
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.2
traitlets==5.14.3
trimesh==4.4.9
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
wcwidth==0.2.13
werkzeug==3.0.4
widgetsnbextension==4.0.13
yapf==0.40.2
zipp==3.20.1

